requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
gunicorn==20.1.0
requests
python-dotenv
qdrant_client
pymupdf
langchain-openai
langchain